fix and move s3 fixtures
